# [FlexBuffers][Java] WIP: Performance improvements on FlexBuffers Java