A series of micro-optimizations to improve FlexBuffer read performance.
It is not the final solution (and testing), but I decided to share already
in order to avoid unused work in case of changes.

1: create a ReadBuf interface to avoid extra work on ByteBuffer.
It seems ByteBuffer has a lot of overhead due to checking bounds
twice (one time on ByteBuffer, second on backing byte[]). Also, there
is a lot of Buffer implementations from different libraries, like Netty
and Okio. So we are creating an interface to be able to plug different
buffer strategies.

The naked array implementation ArrayReadBuf provides a performance Buffer
over regular ByteBuffers due to the fact, ByteBuffers mostly check buffer
bounds twice.

2: cache key position on Map objects.
Read the real position of each key of a map can become very expensive, because
due to binary search, `indirect()` is called several times. Caching this data
is relatively cheap (array of ints with the size of n. of keys). And provides a very good performance boost.

3: Avoid temporary allocations on Map.binarySearch().
Original implementation naively allocates a KeyVector than perform a binary search over it.
We are using the already cache positions to perform the key comparison without that additional
allocation.

4: Optimize String vs buffer comparison for Keys
The original implementation we convert Key to string and do a comparison against the tested key.

We change the implementation to perform a check codepoint by codepoint, instead
of allocating the whole string. So avoid reducing most of the String allocations during
binarySearch().

That PR is not ready and is meant to get feedback. Current benchmark on the matter is the following:
```
benchmark:    45,563,192 ns AsteroidTest.deserializeFlexBuffer
benchmark:     8,699,808 ns AsteroidTest.deserializeFlexBufferReadBuf
benchmark:    37,635,615 ns AsteroidTest.deserializeGson
benchmark:    32,069,347 ns AsteroidTest.deserializeMoshi
```

AsteroidTest.deserializeFlexBufferReadBuf is the proposed version and AsteroidTest.deserializeFlexBuffer
is the original.

I added Moshi and Gson, two most used Json libraries android to use as a comparison.

for more details about the benchmark, check:
https://github.com/paulovap/android-serial-benchmark