@aardappel 2019-10-17
Wow, that's indeed some speedup! How representative is this data of a typical JSON-style data workload?
Do you have any way of telling which of your changes contribute the most to this speedup?

Caching large amounts of values is not a great idea in my mind, since the whole point of the FlatBuffers/FlexBuffers design is "reading in place" such that you don't need additional storage. If we are going to allocate large amounts of cached data, it feels like we might as well have unpacked the data in the first place.

I chose ByteBuffer originally for FlatBuffers as it seemed to do exactly what we needed, but really, for FlexBuffers, maybe it makes more sense to use an array directly? Though I guess your ArrayReadBuf is essentially that?