@paulovap 2019-10-18
> Wow, that's indeed some speedup! How representative is this data of a typical JSON-style data workload?

I can only speak for the company I work for, on the client <-> server context.

The objects(asteroids) size would be an small-to-medium sized object for majority of endpoints. Although, we have some endpoints sending a array of objects with 50+ fields. Of course, we don't send 1000 elements as the benchmark, but the idea is to be able to distinguish the impact better.

If you have a dataset to suggest, I can incorporate on the benchmark project.  
> Do you have any way of telling which of your changes contribute the most to this speedup?

Yes. I did a quick re-run again:

All Optimizations:
`7,663,423 ns AsteroidTest.deserializeFlexBufferReadBuf`
Without Map keys indirect cache
`8,326,923 ns AsteroidTest.deserializeFlexBufferReadBuf`
Without binarySearch Optimizations (allocate KeyVector and allocating strings for comparison)
`36,495,193 ns AsteroidTest.deserializeFlexBufferReadBuf`
Without ReadBuf (original implementation):
`42,734,923 ns AsteroidTest.deserializeFlexBuffer`

> Caching large amounts of values is not a great idea in my mind, since the whole point of the FlatBuffers/FlexBuffers design is "reading in place" such that you don't need additional storage. If we are going to allocate large amounts of cached data, it feels like we might as well have unpacked the data in the first place.

"Reading in place" in Java does not play as well as in C++. Java implementation random access on the `byte[]` is bound checked, making it slow. It is even worse on `ByteBuffer`. I thought that caching real position of the keys in `Map` would be a good speedup, because accessing elements in the map is the most expensive operation by far. Also, you only pay the price of cache if you request access to the map. As in `myRef.asMap()`. If you feel is not worth it, I can remove from the PR.

> I chose ByteBuffer originally for FlatBuffers as it seemed to do exactly what we needed, but really, for FlexBuffers, maybe it makes more sense to use an array directly? Though I guess your ArrayReadBuf is essentially that?

`DirectByteBuffer` is the only way I know to directly access native(JNI) chunks of memory, so it need to be supported anyway. But on the other hand, if you have a `ByteBuffer` backed by an array, its better to access it directly. 

That is why I added the `ReadBuf` interface, so we could plug anything. And yes, it might not being performant as doing array only. But I don't have a better idea to support both: array and `ByteBuffer` as well as allow any other buffer-based implementation. As  [Netty's](https://netty.io/4.1/api/io/netty/buffer/ByteBuf.html) or [okio's](https://square.github.io/okio/1.x/okio/) buffers.