@aardappel 2019-10-21
I don't know what "average" data looks like.. I just hope we're not making decisions based on numbers that are only indicative of a special case.

So it sounds like the binary search optimization is the big one. The other ones.. I guess they can stay in if they don't cause a lot of allocation.

@incubus370 do you have any opinion?