@incubus370 2019-10-24
Caching the positions of all child еlements in the map is definitely wrong. For example, I want only one value, and there are a lot of elements in the map.