@incubus370 2019-10-24
I suggest introducing optimizations gradually. There is a problem of decoding keys into strings. In the decoder, you need to add a method that decodes the string until a zero character appears. In this case, you do not have to iterate over the buffer twice. The first time is the search for the end of the string, the second is decoding.