@aardappel 2019-10-24
@incubus370 agree doing them one at a time would be good, so we can see better what it effects.

In the case of a large map where you only ever query one key, what does the current code do?