@incubus370 2019-10-25
Decoding keys and reading the positions of child elements are different problems.
The question https://github.com/google/flatbuffers/pull/5568#issuecomment-546015842 was addressed to @paulovap