@incubus370 2019-10-25
For the problem of reading the position of all elements, I can suggest creating two implementations of the map. The first reads in memory the positions of the keys, the second does not.