@paulovap 2019-10-25
> Caching the positions of all child еlements in the map is definitely wrong. For example, I want only one value, and there are a lot of elements in the map.

Hum..maybe this is not entirely necessary, indeed. Even though it will have an visible performance benefits over a highly accessed map. It will not benefit maps that are not accessed much. I will remove from the PR.